Almost worked the first time.  The extended LIBPATH support allowed dav_fs.dll to be loaded before dav.dll in httpd.conf.
However, not all of the testcase code made it into mod_so.c correctly with the result that the pre-existing reference to the php extensions directory was lost when the extended LIBPATH was updated and modphp7 failed to load.

This commit corrects the mod_so.c sources to avoid this failure.